Implement more robust event handling for transaction events
Bad error handling can cause an abort, which itself causes another error due to trying to call `completeError` on the completer.

Another issue is that we could subscribe to the transaction events too late, causing a StateError.
This is solved by storing `completed` immediately after the transaction is created to subscribe to the events ASAP.